added 3nd day examples

+++ src/main/java/com/bilgeadam/bilgeadamfullstack/service/StudentService.java
@@ -1,5 +1,7 @@
 package com.bilgeadam.bilgeadamfullstack.service;
 
+import com.bilgeadam.bilgeadamfullstack.exception.BadRequestException;
+import com.bilgeadam.bilgeadamfullstack.exception.StudentNotFoundException;
 import com.bilgeadam.bilgeadamfullstack.model.Student;
 import com.bilgeadam.bilgeadamfullstack.repository.StudentRepository;
 import lombok.AllArgsConstructor;
@@ -18,10 +20,17 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
+        boolean isExists = studentRepository.selectExistsEmail(student.getEmail());
+        if(isExists){
+            throw new BadRequestException("Email " + student.getEmail() +" taken");
+        }
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
+        if(!studentRepository.existsById(studentId)){
+            throw new StudentNotFoundException("Student with id " + studentId + " does not exists");
+        }
         studentRepository.deleteById(studentId);
     }
 }
